+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Kinetra Changelog
 
-## [0.0.1a04] - 2026-09-14
+## [0.0.1a05] - 2026-09-15
 
 ### Added
 
+- Added `sim` blocks.
+- Added brace block syntax using `{` and `}`.
+- Added `step` statement for declaring simulation iteration count.
+- Added built-in `step_index` variable inside simulation loops.
 - Added `let` variable declaratios.
 - Added variable assignment using `=`.
 - Added variable lookup in expressions.
@@ -12,8 +16,13 @@
 - Added `TOKEN_PRINT`.
 - Added `TOKEN_LET`.
 - Added `TOKEN_ASSIGN`.
+- Added `TOKEN_SIM`.
+- Added `TOKEN_STEP`.
+- Added `TOKEN_LBRACE`.
+- Added `TOKEN_RBRACE`.
 - Added `NODE_ASSIGN`.
 - Added `NODE_PRINT` AST node.
+- Added `NODE_STEP`.
 - Added a simple VM symbol table for numeric variables.
 - Added support for single-statement programs using `print expression;`.
 - Added `//` line-comment support to the lexer.
@@ -26,6 +35,7 @@
 
 ### Changed
 
+- `NODE_SIMULATION_BLOCK` is now used by the parser and VM.
 - Program AST now stores a list of statements.
 - Parser now parses a statement instead of only a raw expression.
 - VM now executes `print` statements directly.